requirements.txt
anthropic>=0.40.0
fpdf2>=2.7.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
langchain-core>=0.2.0
pandas>=2.0.0
pillow>=10.0.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pydantic>=2.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.30.0
twilio>=8.0.0
